pyproject.toml
[project]
name = "vlm-home"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.10.1",
    "fastapi>=0.116.1",
    "matplotlib>=3.10.6",
    "opencv-python>=4.12.0.88",
    "pillow>=11.3.0",
    "pydantic-settings>=2.10.1",
    "python-multipart>=0.0.20",
    "timm>=1.0.19",
    "torch>=2.8.0",
    "transformers>=4.56.0",
    "uvicorn>=0.35.0",
]
